tools to check dates and expiartion of txs
