+++ governance/scripts/bridge/_lib.js
@@ -1,7 +1,10 @@
 const { ethers } = require('hardhat')
-const { ADDRESS_ZERO } = require('@unlock-protocol/hardhat-helpers')
+const { ADDRESS_ZERO, getNetwork } = require('@unlock-protocol/hardhat-helpers')
 const { networks } = require('@unlock-protocol/networks')
 
+/***
+ * CONNEXT logic
+ */
 const xCalledABI = [
   {
     anonymous: false,
@@ -169,7 +172,6 @@ const fetchOriginXCall = async ({ transferIds = [], chainId = 1 }) => {
       }
     }
   `
-  console.log(query)
   const { originTransfers } = await fetchXCall({ query, chainId })
   return originTransfers
 }
@@ -217,7 +219,7 @@ const getSupportedChainsByDomainId = async () => {
     )
 }
 
-const subgraphIds = {
+const connextSubgraphIds = {
   1: `FfTxiY98LJG6zoiAjCXdT34pAmCKDEP8vZRVuC8D5Gf`,
   137: `7mDXK2K6UfkVXiJMhXU8VEFuh7qi2TwdYxeyaRjkmexo`, //plygon
   10: `3115xfkzXPrYzbqDHTiWGtzRDYNXBxs8dyitva6J18jf`, //optimims
@@ -225,7 +227,7 @@ const subgraphIds = {
   100: `6oJrPk9YJEU9rWU4DAizjZdALSccxe5ZahBsTtFaGksU`, //gnosis
 }
 
-const subgraphURL = (chainId) => {
+const connextSubgraphURL = (chainId) => {
   // bnb is hosted version
   if (chainId == 56) {
     return 'https://api.thegraph.com/subgraphs/name/connext/amarok-runtime-v0-bnb'
@@ -234,7 +236,7 @@ const subgraphURL = (chainId) => {
   if (!SUBGRAPH_QUERY_API_KEY) {
     throw new Error(`Missing SUBGRAPH_QUERY_API_KEY env`)
   }
-  const subgraphId = subgraphIds[chainId]
+  const subgraphId = connextSubgraphIds[chainId]
   if (!subgraphId) {
     throw new Error(`Unknown chain id ${chainId}`)
   }
@@ -242,7 +244,7 @@ const subgraphURL = (chainId) => {
 }
 
 const fetchXCall = async ({ query, chainId }) => {
-  const endpoint = subgraphURL(chainId)
+  const endpoint = connextSubgraphURL(chainId)
   const q = await fetch(endpoint, {
     method: 'POST',
     headers: { 'Content-Type': 'application/json' },
@@ -259,9 +261,481 @@ const fetchXCall = async ({ query, chainId }) => {
   return data
 }
 
+/***
+ * Delay Mod Logic
+ */
+
+const delayABI = [
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'previousAvatar',
+        type: 'address',
+      },
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'newAvatar',
+        type: 'address',
+      },
+    ],
+    name: 'AvatarSet',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'guard',
+        type: 'address',
+      },
+    ],
+    name: 'ChangedGuard',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'initiator',
+        type: 'address',
+      },
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'owner',
+        type: 'address',
+      },
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'avatar',
+        type: 'address',
+      },
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'target',
+        type: 'address',
+      },
+    ],
+    name: 'DelaySetup',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'module',
+        type: 'address',
+      },
+    ],
+    name: 'DisabledModule',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: false,
+        internalType: 'address',
+        name: 'module',
+        type: 'address',
+      },
+    ],
+    name: 'EnabledModule',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'previousOwner',
+        type: 'address',
+      },
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'newOwner',
+        type: 'address',
+      },
+    ],
+    name: 'OwnershipTransferred',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'previousTarget',
+        type: 'address',
+      },
+      {
+        indexed: true,
+        internalType: 'address',
+        name: 'newTarget',
+        type: 'address',
+      },
+    ],
+    name: 'TargetSet',
+    type: 'event',
+  },
+  {
+    anonymous: false,
+    inputs: [
+      {
+        indexed: true,
+        internalType: 'uint256',
+        name: 'queueNonce',
+        type: 'uint256',
+      },
+      {
+        indexed: true,
+        internalType: 'bytes32',
+        name: 'txHash',
+        type: 'bytes32',
+      },
+      { indexed: false, internalType: 'address', name: 'to', type: 'address' },
+      {
+        indexed: false,
+        internalType: 'uint256',
+        name: 'value',
+        type: 'uint256',
+      },
+      { indexed: false, internalType: 'bytes', name: 'data', type: 'bytes' },
+      {
+        indexed: false,
+        internalType: 'enum Enum.Operation',
+        name: 'operation',
+        type: 'uint8',
+      },
+    ],
+    name: 'TransactionAdded',
+    type: 'event',
+  },
+  {
+    inputs: [],
+    name: 'avatar',
+    outputs: [{ internalType: 'address', name: '', type: 'address' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      { internalType: 'address', name: 'prevModule', type: 'address' },
+      { internalType: 'address', name: 'module', type: 'address' },
+    ],
+    name: 'disableModule',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'address', name: 'module', type: 'address' }],
+    name: 'enableModule',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      { internalType: 'address', name: 'to', type: 'address' },
+      { internalType: 'uint256', name: 'value', type: 'uint256' },
+      { internalType: 'bytes', name: 'data', type: 'bytes' },
+      { internalType: 'enum Enum.Operation', name: 'operation', type: 'uint8' },
+    ],
+    name: 'execTransactionFromModule',
+    outputs: [{ internalType: 'bool', name: 'success', type: 'bool' }],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      { internalType: 'address', name: 'to', type: 'address' },
+      { internalType: 'uint256', name: 'value', type: 'uint256' },
+      { internalType: 'bytes', name: 'data', type: 'bytes' },
+      { internalType: 'enum Enum.Operation', name: 'operation', type: 'uint8' },
+    ],
+    name: 'execTransactionFromModuleReturnData',
+    outputs: [
+      { internalType: 'bool', name: 'success', type: 'bool' },
+      { internalType: 'bytes', name: 'returnData', type: 'bytes' },
+    ],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [
+      { internalType: 'address', name: 'to', type: 'address' },
+      { internalType: 'uint256', name: 'value', type: 'uint256' },
+      { internalType: 'bytes', name: 'data', type: 'bytes' },
+      { internalType: 'enum Enum.Operation', name: 'operation', type: 'uint8' },
+    ],
+    name: 'executeNextTx',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'getGuard',
+    outputs: [{ internalType: 'address', name: '_guard', type: 'address' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      { internalType: 'address', name: 'start', type: 'address' },
+      { internalType: 'uint256', name: 'pageSize', type: 'uint256' },
+    ],
+    name: 'getModulesPaginated',
+    outputs: [
+      { internalType: 'address[]', name: 'array', type: 'address[]' },
+      { internalType: 'address', name: 'next', type: 'address' },
+    ],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [
+      { internalType: 'address', name: 'to', type: 'address' },
+      { internalType: 'uint256', name: 'value', type: 'uint256' },
+      { internalType: 'bytes', name: 'data', type: 'bytes' },
+      { internalType: 'enum Enum.Operation', name: 'operation', type: 'uint8' },
+    ],
+    name: 'getTransactionHash',
+    outputs: [{ internalType: 'bytes32', name: '', type: 'bytes32' }],
+    stateMutability: 'pure',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'uint256', name: '_nonce', type: 'uint256' }],
+    name: 'getTxCreatedAt',
+    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'uint256', name: '_nonce', type: 'uint256' }],
+    name: 'getTxHash',
+    outputs: [{ internalType: 'bytes32', name: '', type: 'bytes32' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'guard',
+    outputs: [{ internalType: 'address', name: '', type: 'address' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'address', name: '_module', type: 'address' }],
+    name: 'isModuleEnabled',
+    outputs: [{ internalType: 'bool', name: '', type: 'bool' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'owner',
+    outputs: [{ internalType: 'address', name: '', type: 'address' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'queueNonce',
+    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'renounceOwnership',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'address', name: '_avatar', type: 'address' }],
+    name: 'setAvatar',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'address', name: '_guard', type: 'address' }],
+    name: 'setGuard',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'address', name: '_target', type: 'address' }],
+    name: 'setTarget',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'uint256', name: 'cooldown', type: 'uint256' }],
+    name: 'setTxCooldown',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'uint256', name: 'expiration', type: 'uint256' }],
+    name: 'setTxExpiration',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'uint256', name: '_nonce', type: 'uint256' }],
+    name: 'setTxNonce',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'bytes', name: 'initParams', type: 'bytes' }],
+    name: 'setUp',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'skipExpired',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'target',
+    outputs: [{ internalType: 'address', name: '', type: 'address' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'address', name: 'newOwner', type: 'address' }],
+    name: 'transferOwnership',
+    outputs: [],
+    stateMutability: 'nonpayable',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'txCooldown',
+    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+    name: 'txCreatedAt',
+    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'txExpiration',
+    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+    name: 'txHash',
+    outputs: [{ internalType: 'bytes32', name: '', type: 'bytes32' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+  {
+    inputs: [],
+    name: 'txNonce',
+    outputs: [{ internalType: 'uint256', name: '', type: 'uint256' }],
+    stateMutability: 'view',
+    type: 'function',
+  },
+]
+
+const getDelayModule = async (delayModuleAddress) => {
+  // fetch delayMod address from networks package
+  if (!delayModuleAddress) {
+    ;({
+      governanceBridge: {
+        modules: { delayMod: delayModuleAddress },
+      },
+    } = await getNetwork())
+  }
+  const delayMod = await ethers.getContractAt(delayABI, delayModuleAddress)
+
+  console.log(`DelayMod at ${delayModuleAddress}`)
+
+  // get the nonces
+  const currentNonce = await delayMod.txNonce()
+  const queueNonce = await delayMod.queueNonce()
+  console.log(`Nonces - current: ${currentNonce}, next ${queueNonce}`)
+
+  return { delayMod, currentNonce, queueNonce }
+}
+
+const fetchDataFromTx = async ({ txHash }) => {
+  const { interface } = await ethers.getContractAt(delayABI, ADDRESS_ZERO)
+
+  // fetch data from tx
+  const { logs } = await ethers.provider.getTransactionReceipt(txHash)
+  const parsedLogs = logs.map((log) => {
+    try {
+      return interface.parseLog(log)
+    } catch (error) {
+      return {}
+    }
+  })
+  const { args } = parsedLogs.find(({ name }) => name === 'TransactionAdded')
+  return args
+}
+
+const logStatus = (transferId, status) => {
+  const { origin, dest } = status
+  const { explorer, name, id } = networks[dest.chainId]
+  console.log(`To ${name} (${id}) - ${transferId}
+  - origin (${origin.status}) - tx : ${explorer.urls.transaction(
+    origin.transactionHash
+  )}
+  - dest (${dest.status})
+    - executedTransactionHash: ${explorer.urls.transaction(
+      dest.executedTransactionHash
+    )}
+    - reconciledTransactionHash ${explorer.urls.transaction(
+      dest.reconciledTransactionHash
+    )}\n`)
+}
+
 module.exports = {
   getXCalledEvents,
   fetchOriginXCall,
   fetchDestinationXCall,
   getSupportedChainsByDomainId,
+  getDelayModule,
+  fetchDataFromTx,
+  logStatus,
 }
